Hide boost rng behind facade for osx compatibility
Split boost random number generation from the common Caffe singleton and
add a helper function for rng. This resolves a build conflict in OSX
between boost rng and nvcc compilation of cuda code.

Refer to #165 for a full discussion.

Thanks to @satol for suggesting a random number generation facade rather
than a total split of cpp and cu code, which is far more involved.


Former-commit-id: 856f05c70c5e8a284f4a96f7b4ea00ea2f0b49ac [formerly 19bcf2b29bf9e48ff84d18763c6d2b5f41e5bdcd]
Former-commit-id: 84016fade4be73d47265f1509f691859a2663a1c