don't squeeze blob arrays for python
Preserve the non-batch dimensions of blob arrays, even for singletons.

The forward() and backward() helpers take lists of ndarrays instead of a
single ndarray per blob, and lists of ndarrays are likewise returned.

Note that for output the blob array could actually be returned as a
single ndarray instead of a list.


Former-commit-id: 16cd5d53036c07a9c191d8bb0abe09f44d28eb2c [formerly 8af33e8ca2f56a3bef23935990b1c9ed65629918]
Former-commit-id: d6195472da90827abf0741b2c13349c3e40b5268