padding for max pooling
Max pooling pads by -inf if the padding parameter is set.

Padding for pooling, like padding for convolution, can preserve the
dimensions of the bottom at the top. By setting the padding to
floor(kernel_size / 2) the top output is the "same" instead of the
"valid" part of the bottom input.


Former-commit-id: 6303940b82ffdd637ccaed06c6e9e2fed900288d [formerly 0d257e431404ad11df48816b85a94898bc86bf13]
Former-commit-id: a23f028269f6c760d0c4c2bfefaf59b956d5510a