do all caffe_copy() as UVA mem copy, and drop caffe_gpu_copy()
Do all memory copies by `cudaMemcpy` in UVA mode so that the same
`caffe_copy()` interface works for all transfers.

`cudaMemcpy()` is used in lieu of BLAS copies because they do not
understand UVA.

Drop the now unnecessary `caffe_gpu_copy()` since location of the
pointers is now irrelevant to the interface.


Former-commit-id: 68cab74fae5248d79544759fad6e617eaf647ad3 [formerly aaa9f24d369798cfa268f465ec5a1c35257ed500]
Former-commit-id: dba324c8dbd66ad2e8ae1e6a9f98fea8d6f93335