Add net tests for loss_weight.
Check that the loss and gradients throughout the net are appropriately scaled
for a few loss_weight values, assuming a default weight of 1 in the loss layer
only.  Also modify test_gradient_check_util to associate a loss of 2 rather
than 1 with the top blob, so that loss layer tests fail if they don't scale
their diffs.


Former-commit-id: bcc66a12f168e079f9b9bfc6db89de20d47fc108 [formerly 7a3ed9b8edf43895770b63cb4d9f5cacf0dba047]
Former-commit-id: b4cb2a2823478478c310261be695a1ef892ecc57