include Reshape in caffe time
Since we are now calling Reshape in the Forward pass, it's only fair to
include it when timing. Reshape calls should normally be four or so
orders of magnitude faster than Forward calls; this change also makes it
easy to notice a mistake that causes something slow to happen in
Reshape.


Former-commit-id: 16bd2c17aef27f8e39fde8c3296feb31bb4abe9f [formerly 24350a6f404b6f8988f742c91fbc07c1541480b7]
Former-commit-id: 29308b27b37997e3408e9bb8beaf1a85aac3f01c